separated FileTree and TreeItem
