+++ CMakeLists.txt
@@ -4,11 +4,14 @@ project(mkfat VERSION 1.0.0
 		DESCRIPTION "a basic MS-DOS disk image formatter"
 		LANGUAGES CXX)
 
-file(GLOB mkfat_src src/*.cpp)
-file(GLOB mkfat_inc include/*.h)
+file(GLOB mkfat_source src/*.cpp)
+file(GLOB mkfat_include include/*.h)
 
-add_executable(mkfat ${mkfat_src} ${mkfat_inc})
+add_executable(mkfat ${mkfat_source} ${mkfat_include})
 
 # we use std::filesystem, so we need c++17
 target_compile_features(mkfat PUBLIC cxx_std_17)
 target_include_directories(mkfat PRIVATE include)
+
+# make sure all errors are known
+target_compile_options(mkfat PRIVATE "-Wall" "-Wextra" "-Wpedantic")
