+++ include/error.h
@@ -1,6 +1,6 @@
 #ifndef _MKFAT_ERROR
 #define _MKFAT_ERROR
 
-_Noreturn void mkfatError(int exitCode, const char *format, ...);
+void mkfatError [[noreturn]] (int exitCode, const char *format, ...);
 
 #endif
